Update README.md
Prerequisites
Java 8 or later
Maven
An IDE (e.g. IntelliJ IDEA, Eclipse)
How to run the tests
Clone the repository to your local machine
Import the project into your IDE as a Maven project
Navigate to the src/test/java directory and run the Runner class as a JUnit test
Writing tests
The framework uses Cucumber to write tests in Gherkin format. The tests are located in the src/test/resources/features directory. To add a new test, create a new .feature file in this directory.